+++ financial-connections-core/src/main/java/com/stripe/android/financialconnections/ConfigurationInternal.kt
@@ -13,7 +13,8 @@ import java.io.Serializable
 data class FinancialConnectionsSheetConfiguration(
     val financialConnectionsSessionClientSecret: String,
     val publishableKey: String,
-    val stripeAccountId: String? = null
+    val stripeAccountId: String? = null,
+    val preCollectedConsent: FinancialConnectionsPreCollectedConsent? = null,
 ) : Parcelable
 
 /**