+++ financial-connections/src/test/java/com/stripe/android/financialconnections/FinancialConnectionsSheetTest.kt
@@ -24,4 +24,33 @@ class FinancialConnectionsSheetTest {
             )
         )
     }
+
+    @Test
+    fun `present() without preCollectedConsent should launch with a null preCollectedConsent`() {
+        financialConnectionsSheet.present(configuration, preCollectedConsent = null)
+        verify(financialConnectionsSheetLauncher).present(
+            configuration = FinancialConnectionsSheetConfiguration(
+                financialConnectionsSessionClientSecret = ApiKeyFixtures.DEFAULT_FINANCIAL_CONNECTIONS_SESSION_SECRET,
+                publishableKey = ApiKeyFixtures.DEFAULT_PUBLISHABLE_KEY,
+                preCollectedConsent = null,
+            ),
+            elementsSessionContext = null,
+        )
+    }
+
+    @Test
+    fun `present() with preCollectedConsent should launch with the given preCollectedConsent`() {
+        val preCollectedConsent = FinancialConnectionsPreCollectedConsent(consent = "fccons_123")
+
+        financialConnectionsSheet.present(configuration, preCollectedConsent = preCollectedConsent)
+
+        verify(financialConnectionsSheetLauncher).present(
+            configuration = FinancialConnectionsSheetConfiguration(
+                financialConnectionsSessionClientSecret = ApiKeyFixtures.DEFAULT_FINANCIAL_CONNECTIONS_SESSION_SECRET,
+                publishableKey = ApiKeyFixtures.DEFAULT_PUBLISHABLE_KEY,
+                preCollectedConsent = preCollectedConsent,
+            ),
+            elementsSessionContext = null,
+        )
+    }
 }